+++ publisher-sdk/src/main/java/com/criteo/publisher/CriteoBannerView.java
@@ -19,6 +19,7 @@ package com.criteo.publisher;
 import android.content.Context;
 import android.util.Log;
 import android.webkit.WebView;
+import androidx.annotation.Keep;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
@@ -90,7 +91,8 @@ public class CriteoBannerView extends WebView {
     }
   }
 
-  public void displayAd(@NonNull String displayData) {
+  @Keep
+  public void loadAdWithDisplayData(@NonNull String displayData) {
     getOrCreateController().displayAd(displayData);
   }